create seaparate split section, adjust structure and style to accomodate multiple panels
