+++ src/components/seda/SedaLocationPanel.js
@@ -1,0 +1,58 @@
+import React, { useMemo } from 'react'
+import { withRouter } from 'react-router-dom';
+import { compose } from 'redux';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types'
+import LocationDetailsPanel from '../organisms/LocationDetailsPanel';
+
+const SedaLocationPanel = ({
+  active,
+  metric,
+  features,
+  selected,
+  clearActiveLocation,
+}) => {
+  // use memo to store other features
+  const others = useMemo(() => 
+    selected.map(fId => features[fId])
+  , [ selected ])
+  return (
+    <LocationDetailsPanel 
+      feature={active} 
+      others={others}
+      onClose={clearActiveLocation}
+      metric={metric}
+    />
+  )
+}
+
+SedaLocationPanel.propTypes = {
+  active: PropTypes.object,
+  metric: PropTypes.string,
+  features: PropTypes.object,
+  region: PropTypes.string,
+  selected: PropTypes.array,
+  clearActiveLocation: PropTypes.func,
+}
+
+const mapStateToProps = 
+  (
+    { features, active, selected },
+    { match: { params: { metric, region } } }
+  ) => ({
+    active,
+    features,
+    region,
+    metric,
+    selected: selected[region]
+  })
+
+const mapDispatchToProps = (dispatch) => ({
+  clearActiveLocation: () => 
+    dispatch({ type: 'CLEAR_ACTIVE_LOCATION'}),
+})
+
+export default compose(
+  withRouter,
+  connect(mapStateToProps, mapDispatchToProps)
+)(SedaLocationPanel)
