+++ src/components/seda/SedaSearch.js
@@ -3,8 +3,7 @@ import {FlyToInterpolator} from 'react-map-gl';
 import * as ease from 'd3-ease';
 import { compose } from 'redux';
 import { withRouter } from 'react-router-dom';
-import { onViewportChange, onRegionChange } from '../../actions/mapActions';
-import { loadLocation } from '../../actions/featuresActions';
+import { onViewportChange, onRegionChange, loadLocation } from '../../actions';
 import { getRegionFromId } from '../../modules/config';
 import Search from '../molecules/Search';
 import { updateRoute } from '../../modules/router';
