+++ src/constants/lang.js
@@ -223,8 +223,8 @@ const LANG = {
   'LEGEND_HIGH': 'higher ▶',
   'LEGEND_LOW_AVG': '◀ below grade level',
   'LEGEND_HIGH_AVG': 'above grade level ▶',
-  'LEGEND_LOW_GRD': '◀ growth below 1 grade',
-  'LEGEND_HIGH_GRD': 'growth above 1 grade ▶',
+  'LEGEND_LOW_GRD': '◀ below 1 grade / year',
+  'LEGEND_HIGH_GRD': 'above 1 grade / year ▶',
   'LEGEND_LOW_COH': '◀ scores dropping',
   'LEGEND_HIGH_COH': 'scores improving ▶',
   'LEGEND_LOW_SES': '◀ poorer',
