+++ src/css/organisms/panel.css
@@ -1,12 +1,10 @@
 .panel {
+  position:relative;
+  z-index:60;
   width: var(--panel-width);  
   flex-shrink:0;
-  margin-left: calc(-1 * var(--panel-width));
-  transition: margin-left 0.2s ease-in-out;
   background: #fff;
   box-shadow: var(--shadow10);
-  position:relative;
-  z-index:50;
   display: flex;
   flex-direction:column;
   align-items: stretch;
@@ -56,18 +54,3 @@
   max-width: calc(100% - 40px);
 }
 
-.panel.location-panel {
-  transition: transform 0.2s ease-in-out;
-  transform: translateX(-100%);
-  position: fixed;
-  top:104px;
-  left: 0px;
-  bottom: 40px;
-  z-index: 60;
-  width: var(--panel-width);
-  margin-left: 0!important;
-}
-
-.location-panel.panel--on {
-  transform: translateX(0);
-}
