+++ src/css/organisms/summary-group.css
@@ -26,6 +26,7 @@
   flex: 1;
   overflow: visible;
   min-height: unset;
+  background: var(--background);
 }
 
 @media(min-width: 600px) {
